--- a/core/test/utils_test.jl
+++ b/core/test/utils_test.jl
@@ -174,7 +174,7 @@ end
     t0 = 0.0
     u0 = Ribasim.build_state_vector(p)
     du0 = copy(u0)
-    p = Ribasim.build_flow_to_storage(p, u0)
+    p = Ribasim.finalize_parameters(p, u0)
     jac_prototype = Ribasim.get_jac_prototype(du0, u0, p, t0)
 
     # rows, cols, _ = findnz(jac_prototype)
@@ -196,7 +196,7 @@ end
     close(db)
     u0 = Ribasim.build_state_vector(p)
     du0 = copy(u0)
-    p = Ribasim.build_flow_to_storage(p, u0)
+    p = Ribasim.finalize_parameters(p, u0)
     jac_prototype = Ribasim.get_jac_prototype(du0, u0, p, t0)
 
     #! format: off