Fixes https://github.com/Deltares/Ribasim/issues/1900.

### Drainage infiltration MWE

#### Main branch

![plot](https://github.com/user-attachments/assets/81430112-dc19-4766-92bf-b377bb965ad7)


#### This branch

![plot](https://github.com/user-attachments/assets/11402f1a-3442-46d1-9e7d-79aca5a24927)

#### Conclusions

Having zero forcing when zero forcing is expected is now enforced properly. In the same way, I'm pretty sure forcing will now never exceed the given flux, and flow should be non-negative where expected.

Having total forcing when the basin is not in the low storage factor regime is harder to enforce, because there is no exact way to infer whether the basin storage passed through the reduction factor regime in the last timestep. A conservative heuristic might suffice here though. 

### Performance comparisons

#### Main branch
- HWS: 9.837521 seconds
- De Dommel: 51.496409 seconds

#### this branch
- HWS: 9.554516 seconds
- De Dommel: 64.802680 seconds