@SouthEndMusic 2024-10-15
It is indeed a fix for #1900, which is one way in which the flows have to be 'tamed' as discussed in #1897. Even though it is not a global fix, I think this PR is a helpful one, as it enforces properties of the solution which we want to be strictly true for any set of solver settings, e.g.
- Non-negative flows where expected
- Realized user demand should not exceed demand

With the ever increasing norm of the state there is still the possibility that the solution slowly drifts away from the physics in ways we cannot capture in the `step_limiter`. If we want to solve that and keep using the fast solver `QNDF`, we either have to find a way to tame the state size (for instance getting resetting to `0` to work), or document that longer simulations require smaller relative errors.