@SouthEndMusic 2024-10-21
I've added the `Average flow states` commit for the record since I was working on it on this branch, but I'm not going to use it. I feel like it's not worth the added complexity, both conceptually and in the code. My goal now is to fix urgent things with the step limiter, and either stop using the relative error (i.e. hardcoding it to be `0`) or documenting that it should be very small if users note growing inaccuracies for long simulations.